bg: Add relativistic background rendering.
This is done by first computing a new texture respecting the light deflection (and the camera view) and then rendering it to the screen.
On the contrary to the particle rendering, the compute shader starts from every pixels position and gets the color from the input texture by computing the angle of a geodesic at infinity.
This geodesic gets its initial parameters by the pixels position.